+++ subprojects/gdk_rust/gdk_electrum/src/lib.rs
@@ -822,8 +822,10 @@ impl ElectrumSession {
 
     pub fn get_receive_address(&self, opt: &GetAddressOpt) -> Result<AddressPointer, Error> {
         debug!("get_receive_address {:?}", opt);
-        let address =
-            self.get_account(opt.subaccount)?.get_next_address(opt.is_internal.unwrap_or(false))?;
+        let address = self.get_account(opt.subaccount)?.get_next_address(
+            opt.is_internal.unwrap_or(false),
+            opt.ignore_gap_limit.unwrap_or(false),
+        )?;
         debug!("get_address {:?}", address);
         Ok(address)
     }